--- templates/carritov2.html
@@ -1,40 +1,0 @@
-{% extends 'app.html' %}
-
-{% block title %}Carrito de Compras{% endblock %}
-
-{% block content %}
-<div class="max-w-4xl mx-auto p-6">
-    <h1 class="text-3xl font-bold mb-6">Carrito de Compras</h1>
-
-    {% if productos %}
-        <div class="grid gap-4">
-            {% for producto in productos %}
-                <div class="bg-white rounded shadow p-4 flex items-center justify-between">
-                    <div class="flex items-center space-x-4">
-                        <img src="{{ producto.imagen_url }}" alt="{{ producto.nombre }}" class="w-24 h-24 object-cover rounded">
-                        <div>
-                            <h2 class="text-xl font-semibold">{{ producto.nombre }}</h2>
-                            <p class="text-gray-600">{{ producto.descripcion }}</p>
-                            <p class="text-blue-600 font-bold mt-2">${{ producto.precio }}</p>
-                        </div>
-                    </div>
-                    <a href="{{ url_for('cartv2.remove_from_cart', producto_id=producto.id) }}"
-                       class="text-red-500 hover:underline">Eliminar</a>
-                </div>
-            {% endfor %}
-        </div>
-        <div class="mt-6 flex justify-between items-center">
-            <a href="{{ url_for('cartv2.clear_cart') }}"
-               class="bg-red-500 hover:bg-red-600 text-white py-2 px-4 rounded">Vaciar carrito</a>
-            <a href="{{ url_for('cartv2.checkout') }}"
-   class="bg-green-500 hover:bg-green-600 text-white py-2 px-4 rounded">Proceder al pago</a>
-
-        </div>
-    {% else %}
-        <p class="text-gray-600">Tu carrito está vacío.</p>
-    {% endif %}
-
-    <a href="{{ url_for('products.products') }}"
-       class="mt-6 inline-block text-blue-600 hover:underline">← Volver a productos</a>
-</div>
-{% endblock %}
